requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
pymongo==4.3.3
pymongo-inmemory==0.2.4
dnspython==2.3.0
certifi==2022.12.7
python-dotenv==1.0.0
cryptography==39.0.1
bcrypt==4.0.1
gunicorn==20.1.0 